fix(emu): address SonarQube static analysis issues across emulator codebase
Fix pointer type truncation issues by using proper width types (uintptr,
DWORD_PTR, WPARAM) instead of int/ulong for pointer casts. Fix null
pointer dereferences where env->user was accessed after nil check on env.
Fix void function returning value in Irix kproc (also added missing
unlock). Fix use-after-free in Nt/devfs.c directory reading. Fix memory
leaks in Nt/win.c attachscreen error paths. Fix division-by-zero guards
in Nt/win.c and Nt/vlrt.c. Fix garbage value from type-punning in
vlrt.c using union. Fix single-iteration loops using if+goto pattern.
Fix uninitialized struct in devpipe.c using mallocz. Fix null pointer
passed to memmove/strncpy in chan.c and dial.c. Fix devmnt.c print that
should be panic for nil mux. Fix uninitialized stat usage in
devfs-posix.c walk. Remove dead debug code in devdraw.c printmesg. Add
null guard for digest block in devssl.c. Use proper int-to-pointer casts
in devprof.c.

+++ DragonFly/386/include/emu.h
@@ -20,7 +20,7 @@ static __inline Proc *getup(void) {
 	__asm__(	"movl	%%esp, %%eax\n\t"
 			: "=a" (p)
 	);
-	return *(Proc **)((unsigned long)p & ~(KSTACK - 1));
+	return *(Proc **)((uintptr)p & ~(KSTACK - 1));
 };
 
 #define	up	(getup())
